feat: integrate shadowing resolution into existing analyze_with_scope_awareness
- Modify analyze_with_scope_awareness and analyze_content_with_scope_awareness
  to use EnhancedDependencyResolver internally
- Add Vec<ShadowingWarning> to return types for shadowing conflict detection
- Remove standalone analyze_with_enhanced_resolution functions
- Consolidate all individual resolvers into single EnhancedDependencyResolver
- Update all tests to handle new return type signature
- Complete Phase 3.3 integration per Issue #103
