--- crates/core/src/shadowing_aware_dependency_resolver.rs
@@ -1,325 +1,0 @@
-use crate::dependency_resolver::DependencyResolver;
-use crate::models::{scope::SymbolTable, Definition, Dependency, DependencyType, Usage, UsageKind};
-use crate::shadowing_resolver::NameResolutionEngine;
-use tree_sitter::Node;
-
-pub struct ShadowingAwareDependencyResolver {
-    base_resolver: Box<dyn DependencyResolver>,
-    symbol_table: Option<SymbolTable>,
-}
-
-impl ShadowingAwareDependencyResolver {
-    pub fn new(base_resolver: Box<dyn DependencyResolver>) -> Self {
-        Self {
-            base_resolver,
-            symbol_table: None,
-        }
-    }
-
-    pub fn with_symbol_table(mut self, symbol_table: SymbolTable) -> Self {
-        self.symbol_table = Some(symbol_table);
-        self
-    }
-
-    fn resolve_with_shadowing_awareness(
-        &self,
-        usage_node: &Usage,
-        definitions: &[Definition],
-    ) -> Vec<Definition> {
-        if let Some(ref symbol_table) = self.symbol_table {
-            let resolver =
-                NameResolutionEngine::new(symbol_table.scopes.clone(), symbol_table.clone());
-
-            let candidates = resolver.resolve_name(usage_node);
-            if !candidates.is_empty() {
-                if let Some(best_candidate) = resolver.select_best_candidate(&candidates) {
-                    return vec![best_candidate.definition.clone()];
-                }
-            }
-        }
-
-        // Fallback to name-based matching
-        definitions
-            .iter()
-            .filter(|def| def.name == usage_node.name)
-            .cloned()
-            .collect()
-    }
-
-    fn create_dependency_from_definition(
-        &self,
-        usage_node: &Usage,
-        definition: &Definition,
-    ) -> Option<Dependency> {
-        let dependency_type = self.get_dependency_type(usage_node);
-        let context = self.get_context(usage_node);
-
-        Some(Dependency {
-            source_line: usage_node.position.start_line,
-            target_line: definition.position.start_line,
-            symbol: usage_node.name.clone(),
-            dependency_type,
-            context,
-        })
-    }
-
-    fn filter_best_matches(
-        &self,
-        candidates: Vec<Definition>,
-        usage_node: &Usage,
-    ) -> Vec<Definition> {
-        if candidates.is_empty() {
-            return candidates;
-        }
-
-        // Prefer definitions that match the usage context
-        let context_matches: Vec<_> = candidates
-            .iter()
-            .filter(|def| self.is_context_compatible(usage_node, def))
-            .cloned()
-            .collect();
-
-        if !context_matches.is_empty() {
-            context_matches
-        } else {
-            // If no context matches, return the first candidate (closest scope)
-            vec![candidates[0].clone()]
-        }
-    }
-
-    fn is_context_compatible(&self, usage_node: &Usage, definition: &Definition) -> bool {
-        match usage_node.kind {
-            UsageKind::CallExpression => {
-                matches!(
-                    definition.definition_type,
-                    crate::models::DefinitionType::FunctionDefinition
-                        | crate::models::DefinitionType::MethodDefinition
-                        | crate::models::DefinitionType::MacroDefinition
-                )
-            }
-            UsageKind::TypeIdentifier => {
-                matches!(
-                    definition.definition_type,
-                    crate::models::DefinitionType::StructDefinition
-                        | crate::models::DefinitionType::EnumDefinition
-                        | crate::models::DefinitionType::TypeDefinition
-                        | crate::models::DefinitionType::ClassDefinition
-                        | crate::models::DefinitionType::InterfaceDefinition
-                )
-            }
-            UsageKind::FieldExpression => {
-                matches!(
-                    definition.definition_type,
-                    crate::models::DefinitionType::StructFieldDefinition
-                        | crate::models::DefinitionType::PropertyDefinition
-                )
-            }
-            _ => true, // For other kinds, accept any definition type
-        }
-    }
-}
-
-impl DependencyResolver for ShadowingAwareDependencyResolver {
-    fn resolve_dependencies(
-        &self,
-        source_code: &str,
-        root_node: Node,
-        usage_nodes: &[Usage],
-        definitions: &[Definition],
-    ) -> Result<Vec<Dependency>, String> {
-        let mut dependencies = Vec::new();
-
-        for usage_node in usage_nodes {
-            let resolved_dependencies =
-                self.resolve_single_dependency(source_code, root_node, usage_node, definitions);
-            dependencies.extend(resolved_dependencies);
-        }
-
-        Ok(dependencies)
-    }
-
-    fn resolve_single_dependency(
-        &self,
-        source_code: &str,
-        root_node: Node,
-        usage_node: &Usage,
-        definitions: &[Definition],
-    ) -> Vec<Dependency> {
-        // First try shadowing-aware resolution
-        let resolved_definitions = self.resolve_with_shadowing_awareness(usage_node, definitions);
-
-        if !resolved_definitions.is_empty() {
-            let filtered_definitions = self.filter_best_matches(resolved_definitions, usage_node);
-
-            return filtered_definitions
-                .into_iter()
-                .filter_map(|def| self.create_dependency_from_definition(usage_node, &def))
-                .collect();
-        }
-
-        // Fallback to base resolver
-        self.base_resolver.resolve_single_dependency(
-            source_code,
-            root_node,
-            usage_node,
-            definitions,
-        )
-    }
-
-    fn get_dependency_type(&self, usage_node: &Usage) -> DependencyType {
-        self.base_resolver.get_dependency_type(usage_node)
-    }
-
-    fn get_context(&self, usage_node: &Usage) -> Option<String> {
-        self.base_resolver.get_context(usage_node)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::models::{
-        scope::{Accessibility, ScopeType},
-        DefinitionType, Position,
-    };
-
-    struct MockDependencyResolver;
-
-    impl DependencyResolver for MockDependencyResolver {
-        fn resolve_dependencies(
-            &self,
-            _source_code: &str,
-            _root_node: Node,
-            _usage_nodes: &[Usage],
-            _definitions: &[Definition],
-        ) -> Result<Vec<Dependency>, String> {
-            Ok(vec![])
-        }
-
-        fn resolve_single_dependency(
-            &self,
-            _source_code: &str,
-            _root_node: Node,
-            _usage_node: &Usage,
-            _definitions: &[Definition],
-        ) -> Vec<Dependency> {
-            vec![]
-        }
-    }
-
-    fn create_test_position(line: usize, column: usize) -> Position {
-        Position {
-            start_line: line,
-            start_column: column,
-            end_line: line,
-            end_column: column + 1,
-        }
-    }
-
-    fn create_test_definition(name: &str, line: usize, def_type: DefinitionType) -> Definition {
-        Definition {
-            name: name.to_string(),
-            definition_type: def_type,
-            position: create_test_position(line, 1),
-        }
-    }
-
-    fn create_test_usage(name: &str, line: usize, kind: UsageKind) -> Usage {
-        Usage {
-            name: name.to_string(),
-            kind,
-            position: create_test_position(line, 5),
-        }
-    }
-
-    #[test]
-    fn test_shadowing_aware_resolution() {
-        let base_resolver = Box::new(MockDependencyResolver);
-        let mut symbol_table = SymbolTable::new();
-
-        // Create nested scopes
-        let func_scope_id = symbol_table.scopes.create_scope(
-            Some(0),
-            ScopeType::Function,
-            create_test_position(5, 1),
-            create_test_position(15, 1),
-        );
-
-        // Add definitions
-        let outer_def = create_test_definition("variable", 1, DefinitionType::VariableDefinition);
-        let inner_def = create_test_definition("variable", 8, DefinitionType::VariableDefinition);
-
-        symbol_table.add_symbol(
-            "variable".to_string(),
-            outer_def.clone(),
-            0,
-            Accessibility::ScopeLocal,
-            false,
-        );
-        symbol_table.add_symbol(
-            "variable".to_string(),
-            inner_def.clone(),
-            func_scope_id,
-            Accessibility::ScopeLocal,
-            false,
-        );
-
-        let resolver =
-            ShadowingAwareDependencyResolver::new(base_resolver).with_symbol_table(symbol_table);
-
-        let usage = create_test_usage("variable", 10, UsageKind::Identifier);
-        let definitions = vec![outer_def, inner_def.clone()];
-
-        let resolved = resolver.resolve_with_shadowing_awareness(&usage, &definitions);
-
-        // Should prefer inner definition due to shadowing
-        assert!(!resolved.is_empty());
-        // Note: This test may fail with current implementation due to the ScopeTree synchronization issue
-        // but demonstrates the intended behavior
-    }
-
-    #[test]
-    fn test_context_compatibility() {
-        let base_resolver = Box::new(MockDependencyResolver);
-        let resolver = ShadowingAwareDependencyResolver::new(base_resolver);
-
-        // Function call usage should match function definitions
-        let func_usage = create_test_usage("test_func", 5, UsageKind::CallExpression);
-        let func_def = create_test_definition("test_func", 1, DefinitionType::FunctionDefinition);
-        let var_def = create_test_definition("test_func", 2, DefinitionType::VariableDefinition);
-
-        assert!(resolver.is_context_compatible(&func_usage, &func_def));
-        assert!(!resolver.is_context_compatible(&func_usage, &var_def));
-
-        // Type usage should match type definitions
-        let type_usage = create_test_usage("TestStruct", 10, UsageKind::TypeIdentifier);
-        let struct_def = create_test_definition("TestStruct", 3, DefinitionType::StructDefinition);
-        let func_def2 = create_test_definition("TestStruct", 4, DefinitionType::FunctionDefinition);
-
-        assert!(resolver.is_context_compatible(&type_usage, &struct_def));
-        assert!(!resolver.is_context_compatible(&type_usage, &func_def2));
-    }
-
-    #[test]
-    fn test_filter_best_matches() {
-        let base_resolver = Box::new(MockDependencyResolver);
-        let resolver = ShadowingAwareDependencyResolver::new(base_resolver);
-
-        let func_usage = create_test_usage("identifier", 5, UsageKind::CallExpression);
-
-        let candidates = vec![
-            create_test_definition("identifier", 1, DefinitionType::VariableDefinition),
-            create_test_definition("identifier", 2, DefinitionType::FunctionDefinition),
-            create_test_definition("identifier", 3, DefinitionType::ConstDefinition),
-        ];
-
-        let filtered = resolver.filter_best_matches(candidates, &func_usage);
-
-        // Should filter to function definition only
-        assert_eq!(filtered.len(), 1);
-        assert!(matches!(
-            filtered[0].definition_type,
-            DefinitionType::FunctionDefinition
-        ));
-    }
-}
