+++ crates/core/tests/mod.rs
@@ -2,6 +2,5 @@ mod metric_calculator_test;
 mod nested_scope;
 mod rust;
 mod scope;
-mod shadowing;
 mod test_utils;
 mod typescript;
