--- crates/core/tests/shadowing/fixtures/mod.rs
@@ -1,4 +1,0 @@
-// Test fixtures for shadowing resolution tests
-
-pub const COMPLEX_RUST_SHADOWING: &str = include_str!("complex_shadowing.rs");
-pub const COMPLEX_TYPESCRIPT_SHADOWING: &str = include_str!("complex_shadowing.ts");
