--- crates/core/tests/shadowing/mod.rs
@@ -1,3 +1,0 @@
-pub mod fixtures;
-pub mod integration_tests;
-pub mod shadowing_resolver_tests;
